requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bertopic==0.16.4
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
bokeh==3.6.3
bs4==0.0.2
cachetools==5.5.2
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
colorcet==3.1.0
colorlog==6.9.0
colorspacious==1.1.2
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cryptography==44.0.0
cuda-bindings==12.8.0
cuda-python==12.8.0
cudf-cu12==25.2.2
cuml-cu12==25.2.1
cupy-cuda12x==13.4.0
cuvs-cu12==25.2.1
cycler==0.12.1
cymem==2.0.11
dask==2024.12.1
dask-cuda==25.2.0
dask-cudf-cu12==25.2.2
dask-expr==1.1.21
databricks-sdk==0.44.1
datamapplot==0.5.1
datashader==0.17.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
distributed==2024.12.1
distributed-ucxx-cu12==0.42.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
duckdb==1.1.3
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.1.0
Faker==19.13.0
fastjsonschema==2.21.1
fastrlock==0.8.3
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-aiplatform==1.81.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.2
google-cloud-resource-manager==1.14.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.14.0
hdbscan==0.8.40
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaleido==0.2.1
keybert==0.8.5
kiwisolver==1.4.8
langchain==0.3.19
langchain-core==0.3.37
langchain-deepseek==0.1.2
langchain-google-vertexai==2.0.13
langchain-ollama==0.2.3
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langcodes==3.5.0
langsmith==0.3.10
language_data==1.3.0
lazy_loader==0.4
Levenshtein==0.26.1
libcudf-cu12==25.2.2
libcuml-cu12==25.2.1
libcuvs-cu12==25.2.1
libkvikio-cu12==25.2.1
libraft-cu12==25.2.0
libucx-cu12==1.18.0
libucxx-cu12==0.42.0
llvmlite==0.43.0
locket==1.0.0
loguru==0.7.3
lz4==4.4.3
Mako==1.3.9
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.2
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
mlflow==2.20.2
mlflow-skinny==2.20.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multipledispatch==1.0.0
murmurhash==1.0.12
mysql-connector-python==9.1.0
narwhals==1.29.1
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.60.0
numba-cuda==0.2.0
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.20.5
nvidia-nvcomp-cu12==4.2.0.11
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
nvtx==0.2.11
ollama==0.4.7
openai==1.60.2
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
optuna==4.2.1
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
param==2.2.0
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==6.0.0
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyahocorasick==2.1.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyct==0.5.0
pydantic==2.10.6
pydantic_core==2.27.2
PyGithub==2.5.0
Pygments==2.19.1
PyJWT==2.10.1
pylabeladjust==0.1.13
pylibcudf-cu12==25.2.2
pylibraft-cu12==25.2.0
PyNaCl==1.5.0
pynndescent==0.5.13
pynvjitlink-cu12==0.5.0
pynvml==12.0.0
pyparsing==3.2.1
Pyqtree==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
raft-dask-cu12==25.2.0
RapidFuzz==3.11.0
rapids-dask-dependency==25.2.0
ray==2.43.0
rcssmin==1.2.1
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rjsmin==1.2.4
rmm-cu12==25.2.0
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.3.1
setuptools==75.8.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.37
sqlglot==25.34.1
sqlglotrs==0.3.0
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
sympy==1.13.1
tblib==3.0.0
tenacity==9.0.0
tensorboardX==2.6.2.2
terminado==0.18.1
thinc==8.3.4
threadpoolctl==3.5.0
tifffile==2025.2.18
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.0
toolz==1.0.0
torch==2.4.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.0
treelite==4.4.1
triton==3.0.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
ucx-py-cu12==0.42.0
ucxx-cu12==0.42.0
umap-learn==0.5.7
uri-template==1.3.0
urllib3==2.3.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.1
xarray==2025.1.2
XlsxWriter==3.2.2
xyzservices==2025.1.0
yarl==1.18.3
zict==3.0.0
zipp==3.21.0
zstandard==0.23.0